image_description_generator/requirements.txt
Pillow==10.1.0
protobuf==4.25.2
python-dotenv==1.0.1
streamlit==1.31.0
google==3.0.0
google-generativeai==0.3.2
